Don't taint master
