+++ test/e2e/ha/upgrade_core_contrail_services_test.go
@@ -23,7 +23,6 @@ import (
 )
 
 func TestUpgradeCoreContrailServices(t *testing.T) {
-	t.SkipNow()
 	if testing.Short() {
 		t.Skip("it is a long test")
 	}
